[trello-106]: Splitted userpanel views into smaller files
